IDEX-3919 Move model in separate artifact
Removed Link class from model classes
Removed ProjectProblem from model classes
