+++ platform-api/che-core-api-model/src/main/java/org/eclipse/che/api/core/model/machine/MachineConfig.java
