+++ platform-api/che-core-api-project/pom.xml
@@ -83,6 +83,10 @@
         </dependency>
         <dependency>
             <groupId>org.eclipse.che.core</groupId>
+            <artifactId>che-core-api-model</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.eclipse.che.core</groupId>
             <artifactId>che-core-api-vfs</artifactId>
         </dependency>
         <dependency>
