+++ platform-api/che-core-api-project/src/main/java/org/eclipse/che/api/project/shared/dto/ProjectTemplateDescriptor.java
@@ -10,10 +10,10 @@
  *******************************************************************************/
 package org.eclipse.che.api.project.shared.dto;
 
-import org.eclipse.che.api.core.model.workspace.ProjectProblem;
 import org.eclipse.che.api.core.rest.shared.dto.Link;
 import org.eclipse.che.api.machine.shared.dto.CommandDto;
 import org.eclipse.che.api.workspace.shared.dto.ProjectConfigDto;
+import org.eclipse.che.api.workspace.shared.dto.ProjectProblemDto;
 import org.eclipse.che.api.workspace.shared.dto.SourceStorageDto;
 import org.eclipse.che.dto.shared.DTO;
 
@@ -70,12 +70,12 @@ public interface ProjectTemplateDescriptor {
 
     ProjectTemplateDescriptor withModules(List<ProjectConfigDto> modules);
 
-    List<ProjectProblem> getProblems();
+    List<ProjectProblemDto> getProblems();
 
     /** @see #getProblems */
-    void setProblems(List<ProjectProblem> problems);
+    void setProblems(List<ProjectProblemDto> problems);
 
-    ProjectTemplateDescriptor withProblems(List<ProjectProblem> problems);
+    ProjectTemplateDescriptor withProblems(List<ProjectProblemDto> problems);
 
     SourceStorageDto getSource();
