+++ platform-api/che-core-api-workspace/src/main/java/org/eclipse/che/api/workspace/shared/dto/ProjectConfigDto.java
@@ -14,7 +14,6 @@ import com.wordnik.swagger.annotations.ApiModelProperty;
 
 import org.eclipse.che.api.core.factory.FactoryParameter;
 import org.eclipse.che.api.core.model.workspace.ProjectConfig;
-import org.eclipse.che.api.core.model.workspace.ProjectProblem;
 import org.eclipse.che.api.core.rest.shared.dto.Link;
 import org.eclipse.che.api.workspace.shared.dto.util.ProjectConfigUtil;
 import org.eclipse.che.dto.shared.DTO;
@@ -95,7 +94,6 @@ public interface ProjectConfigDto extends ProjectConfig {
 
     ProjectConfigDto withSource(SourceStorageDto source);
 
-    @Override
     @FactoryParameter(obligation = OPTIONAL)
     List<Link> getLinks();
 
@@ -106,12 +104,12 @@ public interface ProjectConfigDto extends ProjectConfig {
     /** Provides information about project errors. If project doesn't have any error this field is empty. */
     @ApiModelProperty(value = "Optional information about project errors. If project doesn't have any error this field is empty",
                       position = 17)
-    List<ProjectProblem> getProblems();
+    List<ProjectProblemDto> getProblems();
 
     /** @see #getProblems */
-    void setProblems(List<ProjectProblem> problems);
+    void setProblems(List<ProjectProblemDto> problems);
 
-    ProjectConfigDto withProblems(List<ProjectProblem> problems);
+    ProjectConfigDto withProblems(List<ProjectProblemDto> problems);
 
     @Override
     @FactoryParameter(obligation = OPTIONAL)
