+++ platform-api/pom.xml
@@ -25,15 +25,16 @@
     <modules>
         <module>che-core-api-dto</module>
         <module>che-core-api-dto-maven-plugin</module>
+        <module>che-core-api-model</module>
         <module>che-core-api-core</module>
         <module>che-core-api-vfs</module>
+        <module>che-core-api-workspace</module>
+        <module>che-core-api-user</module>
         <module>che-core-api-project</module>
+        <module>che-core-api-account</module>
         <module>che-core-api-factory</module>
         <module>che-core-api-machine</module>
         <module>che-core-api-analytics</module>
-        <module>che-core-api-user</module>
-        <module>che-core-api-workspace</module>
-        <module>che-core-api-account</module>
         <module>che-core-api-auth</module>
         <module>che-core-api-git</module>
         <module>che-core-api-ssh</module>
